refactor: Refactored based on the other jolt operations
