+++ jolt-core/src/main/java/io/joltcommunity/jolt/enrich/EnrichrExecutionMode.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2013 Bazaarvoice, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.joltcommunity.jolt.enrich;
+
+import io.joltcommunity.jolt.exception.SpecException;
+
+public enum EnrichrExecutionMode {
+    SYNC,
+    ASYNC;
+
+    private static final String EXECUTION_MODE_KEY = "executionMode";
+
+    public static EnrichrExecutionMode fromSpec( Object rawValue ) {
+        if ( rawValue == null ) {
+            return SYNC;
+        }
+        if ( ! ( rawValue instanceof String ) ) {
+            throw new SpecException( "Enrichr optional '" + EXECUTION_MODE_KEY + "' must be a String when provided." );
+        }
+
+        String normalizedValue = ( (String) rawValue ).trim();
+        if ( normalizedValue.isEmpty() ) {
+            throw new SpecException( "Enrichr optional '" + EXECUTION_MODE_KEY + "' must not be blank when provided." );
+        }
+        if ( "sync".equalsIgnoreCase( normalizedValue ) ) {
+            return SYNC;
+        }
+        if ( "async".equalsIgnoreCase( normalizedValue ) ) {
+            return ASYNC;
+        }
+
+        throw new SpecException( "Enrichr optional '" + EXECUTION_MODE_KEY + "' must be either 'sync' or 'async'." );
+    }
+}
